+++ test/unit/v0/test_session.py
@@ -26,28 +26,28 @@ class TestB2Session(TestBase):
         self.api.account_info = self.account_info
 
         self.raw_api = mock.MagicMock()
-        self.raw_api.get_file_info.__name__ = 'get_file_info'
-        self.raw_api.get_file_info.side_effect = ['ok']
+        self.raw_api.get_file_info_by_id.__name__ = 'get_file_info_by_id'
+        self.raw_api.get_file_info_by_id.side_effect = ['ok']
 
         self.session = B2Session(self.account_info, raw_api=self.raw_api)
 
     def test_works_first_time(self):
-        self.assertEqual('ok', self.session.get_file_info(None))
+        self.assertEqual('ok', self.session.get_file_info_by_id(None))
 
     def test_works_second_time(self):
-        self.raw_api.get_file_info.side_effect = [
+        self.raw_api.get_file_info_by_id.side_effect = [
             InvalidAuthToken('message', 'code'),
             'ok',
         ]
-        self.assertEqual('ok', self.session.get_file_info(None))
+        self.assertEqual('ok', self.session.get_file_info_by_id(None))
 
     def test_fails_second_time(self):
-        self.raw_api.get_file_info.side_effect = [
+        self.raw_api.get_file_info_by_id.side_effect = [
             InvalidAuthToken('message', 'code'),
             InvalidAuthToken('message', 'code'),
         ]
         with self.assertRaises(InvalidAuthToken):
-            self.session.get_file_info(None)
+            self.session.get_file_info_by_id(None)
 
     def test_app_key_info_no_info(self):
         self.account_info.get_allowed.return_value = dict(
@@ -56,11 +56,11 @@ class TestB2Session(TestBase):
             capabilities=ALL_CAPABILITIES,
             namePrefix=None,
         )
-        self.raw_api.get_file_info.side_effect = Unauthorized('no_go', 'code')
+        self.raw_api.get_file_info_by_id.side_effect = Unauthorized('no_go', 'code')
         with self.assertRaisesRegexp(
             Unauthorized, r'no_go for application key with no restrictions \(code\)'
         ):
-            self.session.get_file_info(None)
+            self.session.get_file_info_by_id(None)
 
     def test_app_key_info_no_info_no_message(self):
         self.account_info.get_allowed.return_value = dict(
@@ -69,11 +69,11 @@ class TestB2Session(TestBase):
             capabilities=ALL_CAPABILITIES,
             namePrefix=None,
         )
-        self.raw_api.get_file_info.side_effect = Unauthorized('', 'code')
+        self.raw_api.get_file_info_by_id.side_effect = Unauthorized('', 'code')
         with self.assertRaisesRegexp(
             Unauthorized, r'unauthorized for application key with no restrictions \(code\)'
         ):
-            self.session.get_file_info(None)
+            self.session.get_file_info_by_id(None)
 
     def test_app_key_info_all_info(self):
         self.account_info.get_allowed.return_value = dict(
@@ -82,9 +82,9 @@ class TestB2Session(TestBase):
             capabilities=['readFiles'],
             namePrefix='prefix/',
         )
-        self.raw_api.get_file_info.side_effect = Unauthorized('no_go', 'code')
+        self.raw_api.get_file_info_by_id.side_effect = Unauthorized('no_go', 'code')
         with self.assertRaisesRegexp(
             Unauthorized,
             r"no_go for application key with capabilities 'readFiles', restricted to bucket 'my-bucket', restricted to files that start with 'prefix/' \(code\)"
         ):
-            self.session.get_file_info(None)
+            self.session.get_file_info_by_id(None)