+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Add custom exception for `403 transaction_cap_exceeded`
+
 ## [1.2.0] - 2020-11-03
 
 ### Added