+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Add custom exception for `403 transaction_cap_exceeded`
 
+### Changed
+* Fix missing import in the synchronization example
+
 ## [1.2.0] - 2020-11-03
 
 ### Added